refactor: separate view files
