+++ src/views/__init__.py
@@ -1,9 +1,7 @@
-"""
-Views package for yt-dlp Convenient GUI.
-
-This package contains all user interface components and GUI elements.
-"""
-
+from .window_view import WindowMixin
+from .download_tab import DownloadTabMixin
+from .progress_tab import ProgressMixin
+from .event_handlers_view import EventHandlersMixin
+from .refresh_view import RefreshMixin
+from .files_tab import FilesMixin
 from .main_view import MainApplicationView
-
-__all__ = ['MainApplicationView']
