+++ src/views/common/__init__.py
@@ -1,0 +1,2 @@
+from .pixmap_utils import round_pixmap
+from .table_widgets import CopyableTableWidget
