+++ src/views/download_tab/disclaimer.py
@@ -1,0 +1,7 @@
+class DownloadDisclaimerMixin:
+    """Mixin that creates the disclaimer section."""
+
+    def create_disclaimer(self):
+        """Create the disclaimer text."""
+        # Placeholder — the original was an empty frame
+        pass
