+++ src/views/files_tab/constants.py
@@ -1,0 +1,80 @@
+# Map mutagen tag keys to translation keys (human-readable labels)
+_TAG_LABELS = {
+    # ID3
+    "TIT2": "files.title",
+    "TPE1": "files.artist",
+    "TALB": "tag.album",
+    "TRCK": "tag.track",
+    "TPOS": "tag.disc",
+    "TDRC": "tag.year",
+    "TCON": "tag.genre",
+    "TCOM": "tag.composer",
+    "TPUB": "tag.publisher",
+    "TCOP": "tag.copyright",
+    "TENC": "tag.encoder",
+    "TLEN": "tag.length",
+    "APIC": "tag.cover",
+    "TPE2": "tag.album_artist",
+    # MP4
+    "\xa9nam": "files.title",
+    "\xa9ART": "files.artist",
+    "\xa9alb": "tag.album",
+    "trkn": "tag.track",
+    "disk": "tag.disc",
+    "\xa9day": "tag.year",
+    "\xa9gen": "tag.genre",
+    "\xa9wrt": "tag.composer",
+    "\xa9pub": "tag.publisher",
+    "\xa9cpy": "tag.copyright",
+    "\xa9too": "tag.encoder",
+    "aART": "tag.album_artist",
+    # Vorbis / Opus
+    "title": "files.title",
+    "artist": "files.artist",
+    "album": "tag.album",
+    "tracknumber": "tag.track",
+    "discnumber": "tag.disc",
+    "date": "tag.year",
+    "genre": "tag.genre",
+    "composer": "tag.composer",
+    "organization": "tag.publisher",
+    "copyright": "tag.copyright",
+    "encoder": "tag.encoder",
+    "duration": "tag.length",
+    "albumartist": "tag.album_artist",
+    "album artist": "tag.album_artist",
+    "track": "tag.track",
+    "disc": "tag.disc",
+    "year": "tag.year",
+    "length": "tag.length",
+    "publisher": "tag.publisher",
+    "cover": "tag.cover",
+    "lyrics": "files.lyrics",
+    "language": "tag.language",
+    "originaldate": "tag.original_date",
+    "tracktotal": "tag.track_total",
+    # ID3 additional
+    "COMM": "tag.comment",
+    "TDRL": "tag.release_date",
+    "TSSE": "tag.encoding",
+    "TXXX": "tag.user_text",
+    "USLT": "files.lyrics",
+}
+
+_FIELD_KEYS = {
+    'title':      ['TIT2', '\xa9nam', 'title'],
+    'artist':     ['TPE1', '\xa9ART', 'artist'],
+    'album':      ['TALB', '\xa9alb', 'album'],
+    'year':       ['TDRC', '\xa9day', 'date', 'year'],
+    'tracknumber':['TRCK', 'trkn', 'tracknumber'],
+    'genre':      ['TCON', '\xa9gen', 'genre'],
+}
+
+_FIELD_LABELS = {
+    'title':      'files.title',
+    'artist':     'files.artist',
+    'album':      'tag.album',
+    'year':       'tag.year',
+    'tracknumber':'tag.track',
+    'genre':      'tag.genre',
+}
