+++ src/views/progress_tab/__init__.py
@@ -1,0 +1,11 @@
+from .layout import ProgressLayoutMixin
+from .setup import ProgressSetupMixin
+from .eta import ProgressETAMixin
+from .updates import ProgressUpdatesMixin
+from .fetching import ProgressFetchingMixin
+from .tables import ProgressTablesMixin
+
+
+class ProgressMixin(ProgressSetupMixin, ProgressETAMixin,
+                   ProgressUpdatesMixin, ProgressFetchingMixin, ProgressTablesMixin):
+    pass
