--- src/views/widgets_view.py
@@ -1,470 +1,0 @@
-"""
-Widget creation mixin for the main application view.
-
-Contains all create_* methods and related show/hide helpers for
-URL input, path input, format selection, playlist options,
-normalize options, enrich metadata, convert button, and disclaimer.
-"""
-import re
-
-from PySide6.QtWidgets import (
-    QHBoxLayout, QVBoxLayout, QLabel, QLineEdit, QPushButton,
-    QRadioButton, QCheckBox, QComboBox, QButtonGroup, QSpinBox,
-    QSizePolicy, QGroupBox, QMessageBox
-)
-from PySide6.QtGui import QFont, QIcon
-from PySide6.QtCore import Qt, QSize
-
-from config import DEFAULT_BITRATES, DEFAULT_QUALITIES, DOWNLOAD_ICON_PATH, INFO_ICON_PATH
-from utils import settings_manager
-from utils.i18n_utils import t
-
-
-class WidgetsMixin:
-    """Mixin that provides widget creation methods for MainApplicationView."""
-
-    def setup_widgets(self):
-        """Create and layout all GUI widgets."""
-        self.settings_box = QGroupBox(t("settings.group_title"))
-        self.settings_layout = QVBoxLayout(self.settings_box)
-        self.settings_layout.setContentsMargins(5, 10, 5, 10)
-        self.settings_layout.setSpacing(8)
-
-        self.create_url_input()
-        self.create_path_input()
-        self.create_format_selection()
-        self.create_playlist_selection()
-        self.create_options_selection()
-        self.settings_layout.addStretch()
-        self.create_convert_button()
-        self.settings_layout.addStretch()
-        self.create_disclaimer()
-
-        self.input_layout.addWidget(self.settings_box)
-        self.adjust_window_size(margin=1.0)
-        self.setMinimumWidth(self.width() + 20)
-
-    def set_pre_button_spacer_collapsed(self, collapsed: bool):
-        """No-op — spacers are no longer manipulated in the side-by-side layout."""
-
-    def set_post_button_spacer_collapsed(self, collapsed: bool):
-        """No-op — spacers are no longer manipulated in the side-by-side layout."""
-
-    # ------------------------------------------------------------------
-    # URL input
-    # ------------------------------------------------------------------
-
-    def create_url_input(self):
-        """Create URL input field with built-in clear button."""
-        url_layout = QHBoxLayout()
-        url_layout.setContentsMargins(5, 10, 5, 0)
-
-        self.url_entry = QLineEdit()
-        self.url_entry.setPlaceholderText(t("url.placeholder"))
-        self.url_entry.setClearButtonEnabled(True)
-        url_layout.addWidget(self.url_entry, 1)
-
-        self.settings_layout.addLayout(url_layout)
-
-    # ------------------------------------------------------------------
-    # Path input
-    # ------------------------------------------------------------------
-
-    def create_path_input(self):
-        """Create path input with integrated browse icon."""
-        path_layout = QHBoxLayout()
-        path_layout.setContentsMargins(5, 5, 5, 5)
-
-        self.path_entry = QLineEdit()
-        self.path_entry.setReadOnly(True)
-        self.path_entry.setCursor(Qt.PointingHandCursor)
-
-        last_directory = settings_manager.get_last_download_directory()
-        if last_directory:
-            self.path_entry.setText(last_directory)
-        else:
-            self.path_entry.setPlaceholderText(t("path.placeholder"))
-
-        self.path_entry.mousePressEvent = lambda e: self._on_browse_click()
-
-        path_layout.addWidget(self.path_entry, 1)
-
-        self.settings_layout.addLayout(path_layout)
-
-    # ------------------------------------------------------------------
-    # Format selection
-    # ------------------------------------------------------------------
-
-    # Allowed template variables for validation
-    _TEMPLATE_VARIABLES = {"title", "tracknumber", "artist", "album",
-                           "Y", "y", "m", "d", "H", "M", "S", "B", "b"}
-    _TEMPLATE_VAR_RE = re.compile(r'\{(\w+)\}')
-
-    _TEMPLATE_PRESETS = [
-        ("{artist} - {title}",          "format.template_preset_default"),
-        ("{Y}-{m}-{d} - {artist} - {title}", "format.template_preset_date_artist"),
-        ("{Y}{m}{d}_{H}{M}{S}_{title}", "format.template_preset_date_time"),
-        ("{tracknumber} - {artist} - {title}", "format.template_preset_track_artist"),
-        ("{artist}/{album}/{artist} - {title}", "format.template_preset_artist_album"),
-        ("{artist}/{album}/{tracknumber} - {title}", "format.template_preset_album_track"),
-        ("{title}",                      "format.template_preset_simple"),
-    ]
-
-    def create_format_selection(self):
-        """Create file format selection widgets."""
-        self.format_box = QGroupBox(t("format.group_title"))
-        format_box_layout = QVBoxLayout(self.format_box)
-        format_box_layout.setSpacing(6)
-
-        # Row 1: format radios + quality
-        format_row = QHBoxLayout()
-        self.format_group = QButtonGroup(self)
-        self.mp3_radio = QRadioButton(t("format.mp3"))
-        self.mp3_radio.setCursor(Qt.PointingHandCursor)
-        self.mp4_radio = QRadioButton(t("format.mp4"))
-        self.mp4_radio.setCursor(Qt.PointingHandCursor)
-        self.opus_radio = QRadioButton(t("format.opus"))
-        self.opus_radio.setCursor(Qt.PointingHandCursor)
-
-        self.format_group.addButton(self.mp3_radio, 1)
-        self.format_group.addButton(self.mp4_radio, 2)
-        self.format_group.addButton(self.opus_radio, 3)
-
-        format_row.addWidget(self.mp3_radio)
-        format_row.addWidget(self.mp4_radio)
-        format_row.addWidget(self.opus_radio)
-
-        # Quality label + combo box
-        self.quality_label = QLabel(t("quality.label"))
-        self.quality_menu = QComboBox()
-        format_row.addWidget(self.quality_label)
-        format_row.addWidget(self.quality_menu)
-        format_row.addStretch()
-        format_box_layout.addLayout(format_row)
-
-        # Row 2: template input + preset dropdown + info button
-        template_row = QHBoxLayout()
-        self.template_entry = QLineEdit()
-        self.template_entry.setPlaceholderText(t("format.template_placeholder"))
-        if self._output_template_var:
-            self.template_entry.setText(self._output_template_var)
-        self.template_entry.textChanged.connect(self._on_template_text_changed)
-
-        self.template_presets = QComboBox()
-        self.template_presets.setCursor(Qt.PointingHandCursor)
-        for template_val, label_key in self._TEMPLATE_PRESETS:
-            self.template_presets.addItem(t(label_key), template_val)
-        self.template_presets.addItem(t("format.template_preset_custom"), None)
-        # Set initial dropdown selection based on saved template
-        current_template = self.template_entry.text().strip()
-        idx = self.template_presets.findData(current_template)
-        self.template_presets.setCurrentIndex(idx if idx >= 0 else self.template_presets.count() - 1)
-        self.template_presets.currentIndexChanged.connect(self._on_template_preset_changed)
-
-        self.template_info_btn = QPushButton()
-        self.template_info_btn.setIcon(QIcon(INFO_ICON_PATH))
-        self.template_info_btn.setIconSize(QSize(14, 14))
-        self.template_info_btn.setFlat(True)
-        self.template_info_btn.setCursor(Qt.PointingHandCursor)
-        self.template_info_btn.setFixedSize(20, 20)
-        self.template_info_btn.clicked.connect(
-            lambda: QMessageBox.information(self, t("format.template_info_title"), t("format.template_info_text"))
-        )
-
-        template_row.addWidget(self.template_entry, 1)
-        template_row.addWidget(self.template_presets)
-        template_row.addWidget(self.template_info_btn)
-        format_box_layout.addLayout(template_row)
-
-        format_wrapper = QHBoxLayout()
-        format_wrapper.setContentsMargins(5, 0, 5, 0)
-        format_wrapper.addWidget(self.format_box)
-        self.settings_layout.addLayout(format_wrapper)
-
-        # Set initial format from preferences
-        fmt = self._format_var
-        if fmt == 1:
-            self.mp3_radio.setChecked(True)
-            self._populate_bitrate_menu()
-        elif fmt == 2:
-            self.mp4_radio.setChecked(True)
-            self._populate_quality_menu()
-        elif fmt == 3:
-            self.opus_radio.setChecked(True)
-            self._populate_bitrate_menu()
-
-        # Initial validation of template field
-        self._validate_template_visual(self.template_entry.text())
-
-        # Connect signals
-        self.mp3_radio.toggled.connect(lambda checked: self._on_mp3_selected() if checked else None)
-        self.mp4_radio.toggled.connect(lambda checked: self._on_mp4_selected() if checked else None)
-        self.opus_radio.toggled.connect(lambda checked: self._on_opus_selected() if checked else None)
-        self.quality_menu.currentIndexChanged.connect(self._on_quality_or_bitrate_changed)
-
-    def _validate_template_visual(self, text: str):
-        """Apply visual feedback on the template entry: border colour for invalid variables."""
-        if not text.strip():
-            self.template_entry.setStyleSheet("")
-            return
-        invalid = self._get_invalid_template_vars(text)
-        if invalid:
-            self.template_entry.setStyleSheet(
-                "QLineEdit { border: 1px solid #e06060; }"
-            )
-        else:
-            self.template_entry.setStyleSheet("")
-
-    @classmethod
-    def _get_invalid_template_vars(cls, template: str) -> set:
-        """Return the set of variable names in *template* that are not allowed."""
-        used = set(re.findall(cls._TEMPLATE_VAR_RE, template))
-        return used - cls._TEMPLATE_VARIABLES
-
-    @staticmethod
-    def _translate_quality_item(item: str) -> str:
-        """Return the translated display label for a quality/bitrate item."""
-        if item == "Best":
-            return t("quality.best")
-        # "Max 128Kbps" → t("quality.max", value="128Kbps")
-        if item.startswith("Max "):
-            return t("quality.max", value=item[4:])
-        return item
-
-    def _populate_bitrate_menu(self):
-        """Populate quality_menu with bitrate options."""
-        self.quality_menu.blockSignals(True)
-        self.quality_menu.clear()
-        for item in DEFAULT_BITRATES:
-            self.quality_menu.addItem(self._translate_quality_item(item), item)
-        active = self._opus_bitrate_var if self.opus_radio.isChecked() else self._mp3_bitrate_var
-        idx = self.quality_menu.findData(active)
-        if idx >= 0:
-            self.quality_menu.setCurrentIndex(idx)
-        self.quality_menu.blockSignals(False)
-
-    def _populate_quality_menu(self):
-        """Populate quality_menu with quality options."""
-        self.quality_menu.blockSignals(True)
-        self.quality_menu.clear()
-        for item in DEFAULT_QUALITIES:
-            self.quality_menu.addItem(self._translate_quality_item(item), item)
-        idx = self.quality_menu.findData(self._quality_var)
-        if idx >= 0:
-            self.quality_menu.setCurrentIndex(idx)
-        self.quality_menu.blockSignals(False)
-
-    def switch_to_quality_menu(self):
-        """Switch from bitrate to quality menu (MP4)."""
-        self._populate_quality_menu()
-
-    def switch_to_bitrate_menu(self):
-        """Switch from quality to bitrate menu (MP3/Opus)."""
-        self._populate_bitrate_menu()
-
-    # ------------------------------------------------------------------
-    # Playlist selection
-    # ------------------------------------------------------------------
-
-    def create_playlist_selection(self):
-        """Create playlist selection widgets."""
-        self.playlist_box = QGroupBox(t("playlist.group_title"))
-        self.playlist_layout = QHBoxLayout(self.playlist_box)
-
-        self.playlist_group = QButtonGroup(self)
-        self.no_playlist_radio = QRadioButton(t("playlist.no"))
-        self.no_playlist_radio.setCursor(Qt.PointingHandCursor)
-        self.yes_playlist_radio = QRadioButton(t("playlist.yes"))
-        self.yes_playlist_radio.setCursor(Qt.PointingHandCursor)
-
-        self.playlist_group.addButton(self.no_playlist_radio, 1)
-        self.playlist_group.addButton(self.yes_playlist_radio, 0)
-
-        self.playlist_layout.addWidget(self.no_playlist_radio)
-        self.playlist_layout.addWidget(self.yes_playlist_radio)
-
-        # Playlist range widgets (hidden by default)
-        self.playlist_from_label = QLabel(t("playlist.from_video"))
-        self.playlist_start_entry = QSpinBox()
-        self.playlist_start_entry.setRange(1, 9999)
-        self.playlist_start_entry.setValue(self._playlist_start_var)
-        self.playlist_start_entry.setFixedWidth(70)
-        self.playlist_to_label = QLabel(t("playlist.to"))
-        self.playlist_end_entry = QSpinBox()
-        self.playlist_end_entry.setRange(1, 9999)
-        self.playlist_end_entry.setValue(self._playlist_end_var)
-        self.playlist_end_entry.setFixedWidth(70)
-
-        self.playlist_layout.addWidget(self.playlist_from_label)
-        self.playlist_layout.addWidget(self.playlist_start_entry)
-        self.playlist_layout.addWidget(self.playlist_to_label)
-        self.playlist_layout.addWidget(self.playlist_end_entry)
-        self.playlist_layout.addStretch()
-
-        playlist_wrapper = QHBoxLayout()
-        playlist_wrapper.setContentsMargins(5, 0, 5, 0)
-        playlist_wrapper.addWidget(self.playlist_box)
-        self.settings_layout.addLayout(playlist_wrapper)
-
-        # Set initial state from preferences
-        if self._playlist_var == 0:
-            self.yes_playlist_radio.setChecked(True)
-            self._set_playlist_options_visible(True)
-        else:
-            self.no_playlist_radio.setChecked(True)
-            self._set_playlist_options_visible(False)
-
-        # Connect signals
-        self.yes_playlist_radio.toggled.connect(
-            lambda checked: self._on_playlist_selected() if checked else None
-        )
-        self.no_playlist_radio.toggled.connect(
-            lambda checked: self._on_no_playlist_selected() if checked else None
-        )
-        self.playlist_start_entry.valueChanged.connect(self._on_playlist_range_changed)
-        self.playlist_end_entry.valueChanged.connect(self._on_playlist_range_changed)
-
-    def _set_playlist_options_visible(self, visible: bool):
-        """Show or hide playlist range widgets."""
-        self.playlist_from_label.setVisible(visible)
-        self.playlist_start_entry.setVisible(visible)
-        self.playlist_to_label.setVisible(visible)
-        self.playlist_end_entry.setVisible(visible)
-
-    def show_playlist_options(self):
-        """Show playlist range selection widgets."""
-        self._set_playlist_options_visible(True)
-
-    def hide_playlist_options(self):
-        """Hide playlist range selection widgets."""
-        self._set_playlist_options_visible(False)
-
-    # ------------------------------------------------------------------
-    # Normalize selection
-    # ------------------------------------------------------------------
-
-    def create_options_selection(self):
-        """Create options fieldset with normalize, enrich and prevent-sleep checkboxes."""
-        self.options_box = QGroupBox(t("extras.group_title"))
-        options_layout = QVBoxLayout(self.options_box)
-
-        # Normalize volume
-        self.normalize_layout = QHBoxLayout()
-
-        self.normalize_check = QCheckBox(t("options.normalize_volume"))
-        self.normalize_check.setCursor(Qt.PointingHandCursor)
-        self.normalize_check.setChecked(self._normalize_var)
-        self.normalize_check.toggled.connect(self._on_normalize_toggled)
-        self.normalize_layout.addWidget(self.normalize_check)
-
-        # Normalize target widgets (hidden by default)
-        self.normalize_target_label = QLabel(t("options.normalize_target"))
-        self.normalize_layout.addWidget(self.normalize_target_label)
-
-        self.normalize_target_entry = QLineEdit()
-        self.normalize_target_entry.setFixedWidth(50)
-        self.normalize_target_entry.setText(str(self._normalize_target_var))
-        self.normalize_layout.addWidget(self.normalize_target_entry)
-
-        self.normalize_info_btn = QPushButton()
-        self.normalize_info_btn.setIcon(QIcon(INFO_ICON_PATH))
-        self.normalize_info_btn.setIconSize(QSize(14, 14))
-        self.normalize_info_btn.setFlat(True)
-        self.normalize_info_btn.setCursor(Qt.PointingHandCursor)
-        self.normalize_info_btn.setFixedSize(20, 20)
-        self.normalize_info_btn.clicked.connect(
-            lambda: QMessageBox.information(self, t("options.normalize_volume"), t("options.normalize_tooltip"))
-        )
-        self.normalize_layout.addWidget(self.normalize_info_btn)
-        self.normalize_layout.addStretch()
-
-        options_layout.addLayout(self.normalize_layout)
-
-        # Set initial visibility
-        visible = self._normalize_var
-        self.normalize_target_label.setVisible(visible)
-        self.normalize_target_entry.setVisible(visible)
-
-        # Enrich metadata
-        enrich_layout = QHBoxLayout()
-
-        self.enrich_check = QCheckBox(t("options.enrich_metadata"))
-        self.enrich_check.setCursor(Qt.PointingHandCursor)
-        self.enrich_check.setChecked(self._enrich_var)
-        self.enrich_check.toggled.connect(self._on_enrich_toggled)
-        enrich_layout.addWidget(self.enrich_check)
-
-        self.enrich_info_btn = QPushButton()
-        self.enrich_info_btn.setIcon(QIcon(INFO_ICON_PATH))
-        self.enrich_info_btn.setIconSize(QSize(14, 14))
-        self.enrich_info_btn.setFlat(True)
-        self.enrich_info_btn.setCursor(Qt.PointingHandCursor)
-        self.enrich_info_btn.setFixedSize(20, 20)
-        self.enrich_info_btn.clicked.connect(
-            lambda: QMessageBox.information(self, t("options.enrich_metadata"), t("options.enrich_tooltip"))
-        )
-        enrich_layout.addWidget(self.enrich_info_btn)
-        enrich_layout.addStretch()
-
-        options_layout.addLayout(enrich_layout)
-
-        # Prevent sleep
-        sleep_layout = QHBoxLayout()
-
-        self.prevent_sleep_check = QCheckBox(t("options.prevent_sleep"))
-        self.prevent_sleep_check.setCursor(Qt.PointingHandCursor)
-        self.prevent_sleep_check.setChecked(self._prevent_sleep_var)
-        self.prevent_sleep_check.toggled.connect(self._on_prevent_sleep_toggled)
-        sleep_layout.addWidget(self.prevent_sleep_check)
-        sleep_layout.addStretch()
-
-        options_layout.addLayout(sleep_layout)
-
-        options_wrapper = QHBoxLayout()
-        options_wrapper.setContentsMargins(5, 0, 5, 0)
-        options_wrapper.addWidget(self.options_box)
-
-        self.settings_layout.addLayout(options_wrapper)
-
-    def show_normalize_input(self):
-        """Show the normalize target LUFS input."""
-        self.normalize_target_label.setVisible(True)
-        self.normalize_target_entry.setVisible(True)
-
-    def hide_normalize_input(self):
-        """Hide the normalize target LUFS input."""
-        self.normalize_target_label.setVisible(False)
-        self.normalize_target_entry.setVisible(False)
-
-    # ------------------------------------------------------------------
-    # Convert button
-    # ------------------------------------------------------------------
-
-    def create_convert_button(self):
-        """Create the main convert button."""
-        self.convert_button = QPushButton(" " + t("button.download"))
-        self.convert_button.setIcon(QIcon(DOWNLOAD_ICON_PATH))
-        self.convert_button.setIconSize(QSize(18, 18))
-        self.convert_button.setCursor(Qt.PointingHandCursor)
-        self.convert_button.setStyleSheet("""
-            QPushButton {
-                background-color: #238a45;
-                color: white;
-                border: none;
-                border-radius: 4px;
-                padding: 8px 16px;
-            }
-            QPushButton:hover { background-color: #449468; }
-            QPushButton:disabled { background-color: palette(mid); color: palette(dark); }
-        """)
-        self.convert_button.clicked.connect(self._on_convert_click)
-        self.settings_layout.addWidget(self.convert_button, alignment=Qt.AlignCenter)
-
-    # ------------------------------------------------------------------
-    # Disclaimer
-    # ------------------------------------------------------------------
-
-    def create_disclaimer(self):
-        """Create the disclaimer text."""
-        # Placeholder — the original was an empty frame
-        pass
